+++ src/main/java/model/metodologia/condiciones/BooleanCondition.java
@@ -1,7 +1,9 @@
 package model.metodologia.condiciones;
 
+import java.math.BigDecimal;
+
 public interface BooleanCondition {
 	//Se compara siempre el valor 1, por ejemplo en 
 	//LessThan si valorUno es menor que el valorDos retorna true
-	public boolean comparar(Integer valorUno, Integer valorDos);
+	public boolean comparar(BigDecimal valorUno, BigDecimal valorDos);
 }
